refactor(web): split thirteenth-salary page into composable + form/actions/bridge-modals subcomponents
Refs #646

Breaks the 739-LOC thirteenth-salary page down the same way as the
prior pages in this series (hora-extra #706, QuickTransactionForm #707,
WalletEntryForm #708, ProfileCompletionModal #709):

- useThirteenthSalaryPage composable — form state, mutations, modal
  state, calculation + save/goal/budget handlers
- ThirteenthSalaryForm — input fields + tooltips + calculate/reset
- ThirteenthSalaryActions — save button, premium CTA, disclaimer
- ThirteenthSalaryBridgeModals — goal + budget modals (goalForm
  shared via provide/inject to avoid vue/no-mutating-props)
- page.vue — thin orchestrator (149 LOC) wiring the composable to
  the three subcomponents, trimming ~75 LOC of dead CSS selectors

No behavior change; coverage and quality-check green.
